GH-2920: Fix unit test
