+++ projects/storefrontlib/src/cms-components/product/product-intro/product-intro.module.ts
@@ -1,20 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import {
-  CmsConfig,
-  CmsModule,
-  ConfigModule,
-  I18nModule,
-} from '@spartacus/core';
-import { OutletModule } from '../../../cms-structure/outlet/index';
+import { CmsConfig, ConfigModule, I18nModule } from '@spartacus/core';
 import { StarRatingModule } from '../../../shared/components/star-rating/star-rating.module';
 import { ProductIntroComponent } from './product-intro.component';
 
 @NgModule({
   imports: [
     CommonModule,
-    CmsModule,
-    OutletModule,
     I18nModule,
     StarRatingModule,
     ConfigModule.withConfig(<CmsConfig>{
