+++ projects/storefrontlib/src/cms-components/product/product-summary/product-summary.component.spec.ts
@@ -1,4 +1,3 @@
-import { Component, Input } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { I18nTestingModule, Product } from '@spartacus/core';
 import { AddToCartModule } from '../../../cms-components/cart/index';
@@ -7,24 +6,6 @@ import { FormComponentsModule } from '../../../shared/components/form-components
 import { ProductSummaryComponent } from '../product-summary/product-summary.component';
 import { Observable, of } from 'rxjs';
 import { CurrentProductService } from '../current-product.service';
-
-@Component({
-  selector: 'cx-star-rating',
-  template: '',
-})
-class MockStarRatingComponent {
-  @Input() rating;
-  @Input() disabled;
-}
-
-@Component({
-  selector: 'cx-page-slot',
-  template: '',
-})
-class MockDynamicSlotComponent {
-  @Input()
-  position: string;
-}
 
 class MockCurrentProductService {
   getProduct(): Observable<Product> {
@@ -39,12 +20,7 @@ describe('ProductSummaryComponent in product', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [AddToCartModule, FormComponentsModule, I18nTestingModule],
-      declarations: [
-        ProductSummaryComponent,
-        OutletDirective,
-        MockStarRatingComponent,
-        MockDynamicSlotComponent,
-      ],
+      declarations: [ProductSummaryComponent, OutletDirective],
       providers: [
         { provide: CurrentProductService, useClass: MockCurrentProductService },
       ],
@@ -58,63 +34,5 @@ describe('ProductSummaryComponent in product', () => {
 
   it('should be created', () => {
     expect(productSummaryComponent).toBeTruthy();
-  });
-
-  describe('clickTabIfInactive to click tabs indicated as inactive', () => {
-    it('should click tab with no classes', () => {
-      const tabElement: HTMLElement = document.createElement('div');
-      spyOn(tabElement, 'click');
-      productSummaryComponent.clickTabIfInactive(tabElement);
-      expect(tabElement.click).toHaveBeenCalled();
-    });
-
-    it('should not click tab with active class', () => {
-      const tabElement: HTMLElement = document.createElement('div');
-      tabElement.classList.add('active');
-      spyOn(tabElement, 'click');
-      productSummaryComponent.clickTabIfInactive(tabElement);
-      expect(tabElement.click).not.toHaveBeenCalled();
-    });
-
-    it('should click tab with toggled classes', () => {
-      const tabElement: HTMLElement = document.createElement('div');
-      tabElement.classList.add('toggled');
-      spyOn(tabElement, 'click');
-      productSummaryComponent.clickTabIfInactive(tabElement);
-      expect(tabElement.click).toHaveBeenCalled();
-    });
-
-    it('should click tab with active and toggled classes', () => {
-      const tab: HTMLElement = document.createElement('div');
-      tab.classList.add('active');
-      tab.classList.add('toggled');
-      spyOn(tab, 'click');
-      productSummaryComponent.clickTabIfInactive(tab);
-      expect(tab.click).toHaveBeenCalled();
-    });
-  });
-
-  describe('getTabByLabel to get tab from tabs component', () => {
-    it('should return correct tab', () => {
-      const tabsComponent: HTMLElement = document.createElement('div');
-      const tab1: HTMLElement = document.createElement('h3');
-      const tab2: HTMLElement = document.createElement('h3');
-      const tab3: HTMLElement = document.createElement('h3');
-
-      tab1.innerText = 'Tab 1';
-      tab2.innerText = 'Tab 2';
-      tab3.innerText = 'Tab 3';
-
-      tabsComponent.appendChild(tab1);
-      tabsComponent.appendChild(tab2);
-      tabsComponent.appendChild(tab3);
-
-      const result = productSummaryComponent.getTabByLabel(
-        'Tab 2',
-        tabsComponent
-      );
-
-      expect(result).toBe(tab2);
-    });
   });
 });
